# Split structural iteration from stage 2 parser (5% overall)