… src/CMakeLists.txt
… src/westmere/stage2_build_tape.h