In a first move towards a "stage 1 sax" for minification/pretty printing, I made a structural_iterator interface that takes care of advance_char() over structurals, and separated that from stage 2.

It buys us 5% overall (and 10% in stage 2), on Kady Lake R. That wasn't my intent at all :)

I'm guessing that, by peeling off `structural_indexes` into a specific member, the compiler now feels justified in enregistering it, and it's used often enough to make a difference.

### With patch

```
jkeiser@JKEISER-THINKPAD:~/simdjson$ make parse && ./parse jsonexamples/twitter.json -n 20000
g++ -msse4.2 -mpclmul  -std=c++17  -pthread -Wall -Wextra -Wshadow -Ibenchmark/linux -O3 -Isrc -Iinclude  -o parse benchmark/parse.cpp src/simdjson.cpp 
perf_event_open: Invalid argument
number of iterations 20000 
                                                     
jsonexamples/twitter.json
=========================
     9867 blocks -     631515 bytes - 55263 structurals (  8.8 %)
special blocks with: utf8      2284 ( 23.1 %) - escape       598 (  6.1 %) - 0 structurals      1287 ( 13.0 %) - 1+ structurals      8581 ( 87.0 %) - 8+ structurals      3272 ( 33.2 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8      1104 ( 11.2 %) - escape       642 (  6.5 %) - 0 structurals       940 (  9.5 %) - 1+ structurals       940 (  9.5 %) - 8+ structurals      2593 ( 26.3 %) - 16+ structurals         0 (  0.0 %)

All Stages
|    Speed        :  23.3989 ns per block ( 96.86%) -   0.3656 ns per byte -   4.1782 ns per structural -    2.735 GB/s
|- Stage 1
|    Speed        :  11.0762 ns per block ( 45.85%) -   0.1731 ns per byte -   1.9778 ns per structural -    5.778 GB/s
|- Stage 2
|    Speed        :  12.1707 ns per block ( 50.38%) -   0.1902 ns per byte -   2.1732 ns per structural -    5.258 GB/s
```

### Without patch

```
jkeiser@JKEISER-THINKPAD:~/simdjson$ make parse && ./parse jsonexamples/twitter.json -n 20000
make: 'parse' is up to date.
perf_event_open: Invalid argument
number of iterations 20000 
                                                     
jsonexamples/twitter.json
=========================
     9867 blocks -     631515 bytes - 55263 structurals (  8.8 %)
special blocks with: utf8      2284 ( 23.1 %) - escape       598 (  6.1 %) - 0 structurals      1287 ( 13.0 %) - 1+ structurals      8581 ( 87.0 %) - 8+ structurals      3272 ( 33.2 %) - 16+ structurals         0 (  0.0 %)
special block flips: utf8      1104 ( 11.2 %) - escape       642 (  6.5 %) - 0 structurals       940 (  9.5 %) - 1+ structurals       940 (  9.5 %) - 8+ structurals      2593 ( 26.3 %) - 16+ structurals         0 (  0.0 %)

All Stages
|    Speed        :  24.5744 ns per block ( 97.59%) -   0.3840 ns per byte -   4.3881 ns per structural -    2.604 GB/s
|- Stage 1
|    Speed        :  11.1573 ns per block ( 44.31%) -   0.1743 ns per byte -   1.9923 ns per structural -    5.736 GB/s
|- Stage 2
|    Speed        :  13.2651 ns per block ( 52.68%) -   0.2073 ns per byte -   2.3687 ns per structural -    4.824 GB/s
```