@jkeiser 2020-03-12
I just pushed a few more improvements intended to shore up the iterator interface (i.e. make more of it private), which push stage 2 a little further:

* Make number, atom and string parsers take a single pointer (&buf[idx]) instead of a pointer and index (buf and idx), to allow a single call to the iterator to retrieve the current structural's buffer position.
* Simplify atom parsing code (same algorithm, a bunch less code)
* Stop copying the buffer when parsing a single atom

I made a version of the patch that didn't have to copy the buffer to parse numbers, but that reversed the (tiny) gains I got from src / buf[idx], so I reverted it for now.