@lemire 2020-03-13
It is definitively faster.