@lemire 2020-03-13
It looks like you are saving half of an instruction per structural character in stage 2.