@lemire 2020-03-13
@jkeiser 

Please merge as early as possible. This is quite nice.